requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.11
python-dotenv>=1.0.0
faiss-cpu>=1.7.4